refactor: dynamic topic discovery for trainer route
- separated the curriculum rendering and related concerns from the router
- created new curriculum_service and curriculum router files
- moved business logic from trainer router to trainer_service
